Backend/requirements.txt
httpx>=0.23.0
pydantic>=1.9.0,<3.0.0
typing-extensions>=4.0.0
cachetools>=4.2.2
zai-sdk